fixed bug where relative time difference differed from most recent version
This fix is for a bug that appeared during testing of COP-9050. Relative time difference shown
in task list was opposite of what was in the most recent version of a task.
Task List would show 'Booked 2 hours before travel', most recent version of a task would show
'Booked 2 hours after travel'.